requirements.txt
Flask==3.0.0
pdfplumber==0.10.3
python-docx==1.1.0
Werkzeug==3.0.1
flask-cors==4.0.0
gunicorn==21.2.0